change mongo driver to 1.4.x; change cluster manager return code to digital; add region field to quota info; fix clusterops error when cluster has multiple master nodes

+++ bcs-services/bcs-cluster-manager/internal/actions/namespace/delete.go
@@ -106,7 +106,7 @@ func (da *DeleteAction) deleteNamespaceFromStore() error {
 	return da.model.DeleteNamespace(da.ctx, da.req.Name, da.req.FederationClusterID)
 }
 
-func (da *DeleteAction) setResp(code uint64, msg string) {
+func (da *DeleteAction) setResp(code uint32, msg string) {
 	da.resp.Code = code
 	da.resp.Message = msg
 	da.resp.Result = (code == types.BcsErrClusterManagerSuccess)
